Area of triangle

+++ 03.Conditionals-Loops/AreaOfTriangle.java
@@ -6,10 +6,10 @@ public class AreaOfTriangle {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter height of the triangle : ");
-        int height = input.nextDouble();
+        double height = input.nextDouble();
 
         System.out.println("Enter base of the triangle : ");
-        int base = input.nextDouble();
+        double base = input.nextDouble();
         
         double area = (height*base)/2;
         System.out.println("Area of the trinagle is : " + area);
